move fontTools.misc.svgPath to fontTools.svgLib
