+++ Lib/fontTools/svgLib/__init__.py
