+++ Lib/fontTools/svgLib/parser.py
