+++ Tests/svgLib/svgLib_test.py
@@ -2,7 +2,7 @@ from __future__ import print_function, absolute_import, division
 
 from fontTools.misc.py23 import *
 from fontTools.pens.recordingPen import RecordingPen
-from fontTools.misc.svgPath import SVGPath
+from fontTools.svgLib import SVGPath
 
 import os
 from tempfile import NamedTemporaryFile
